extract history, move rendering into app.html.app
