--- history/index.js
@@ -1,10 +1,0 @@
-module.exports ={
-  obs: {
-    store: require('./obs/store'),
-  },
-  sync: {
-    push: require('./sync/push'),
-    back: require('./sync/back')
-  }
-}
-
