+++ index.js
@@ -1,7 +1,6 @@
 const ticktack = {
   app: require('./app'),
   blob: require('./blob'),
-  history: require('./history'),
   router: require('./router'),
   styles: require('./styles')
 }
